add fixed exploit test in challenge before test and enable the patch
